# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SEO Agency API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContactFormInputService =
  (typeof ContactFormInputService)[keyof typeof ContactFormInputService];

export const ContactFormInputService = {
  seo: "seo",
  "local-seo": "local-seo",
  "ecommerce-seo": "ecommerce-seo",
  "technical-seo": "technical-seo",
  "link-building": "link-building",
  "content-marketing": "content-marketing",
  other: "other",
} as const;

export interface ContactFormInput {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  email: string;
  /** @maxLength 100 */
  company?: string;
  /** @maxLength 20 */
  phone?: string;
  service: ContactFormInputService;
  /**
   * @minLength 10
   * @maxLength 2000
   */
  message: string;
  /** @maxLength 200 */
  website?: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  company?: string | null;
  phone?: string | null;
  service: string;
  message: string;
  website?: string | null;
  createdAt: string;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  category: string;
  author: string;
  authorRole: string;
  readTime: number;
  publishedAt: string;
  imageUrl?: string | null;
  tags: string[];
}

export type ErrorResponseDetails = { [key: string]: unknown };

export interface ErrorResponse {
  error: string;
  details?: ErrorResponseDetails;
}