# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SEO Agency API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      description: Submits a contact form inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactFormInput"
      responses:
        "201":
          description: Contact submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listContacts
      tags: [contact]
      summary: List all contact submissions
      description: Returns all contact form submissions (admin use)
      responses:
        "200":
          description: List of contact submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSubmission"
...
components:
  schemas:
...
    ContactFormInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        company:
          type: string
          maxLength: 100
        phone:
          type: string
          maxLength: 20
        service:
          type: string
          enum: [seo, local-seo, ecommerce-seo, technical-seo, link-building, content-marketing, other]
        message:
          type: string
          minLength: 10
          maxLength: 2000
        website:
          type: string
          maxLength: 200
      required:
        - name
        - email
        - message
        - service
    ContactSubmission:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        company:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        service:
          type: string
        message:
          type: string
        website:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - service
        - message
        - createdAt
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error